fix velocity clamping!
this clamps velocity by simply looking at the magnitude of the vector.
this is much less error prone than clamping based on what direction
the player is trying to move, and seems to get rid of the square flying
pattern.

I believe all that's left is to define sensible acceleration profiles
so that ships turn more or less as expected.