misc. render perf improvements
the draw many call now borrows the list of quads to draw using a smart
pointer & a refcell.

this will crash at runtime if an entity attempts to mutate memory which
has already been submitted to the render queue. (which is awesome!)

also the vertex buffer should be much smaller now, hoping this improves
memory consumption / bandwidth on the GPU side of things, at least a
little bit.